Test: Move extract_bills_summary function to run_bill_recommendation.py
